requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
bcrypt==4.0.1
psycopg2-binary==2.9.7
flask-cors==4.0.0
email-validator==2.0.0
flask-restx==1.1.0
